requirements.txt
# 审计凭证数据处理与分析系统 - 完整依赖列表

# ========================
# 数据库转换模块依赖
# ========================
pandas>=1.3.0

# ========================
# 自然语言查询系统依赖
# ========================
streamlit>=1.28.0
requests>=2.31.0
openai>=1.0.0  # Compatible with DeepSeek API
python-dotenv>=1.0.0  # Environment variable management

# ========================
# 开发工具依赖（可选）
# ========================
# pytest>=7.0.0          # 测试框架
# black>=23.0.0          # 代码格式化
# flake8>=6.0.0          # 代码检查
# mypy>=1.0.0            # 类型检查